backend/requirements.txt
# Core dependencies
jinja2
tqdm
absl-py
pyyaml
pandas

# LLM APIs
openai>=1.0.0
google-cloud-aiplatform>=1.38.0
anthropic>=0.18.0

# A2A Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Configuration
toml>=0.10.2
python-dotenv>=1.0.0

# Utilities
marko